Create messages table
I created a messages table so that messages can now be stored and retrieved. Created the necessary model/controller methods and routes for messages to be stored in the new table. Did some minor refactoring which involved renaming some variables to match the columns in the DB. Simplified response.js for PUT/POST/DELETE because we really don't need the extra functionality. Moved the chat window from the homepage to the discover page. Cleaned up some minor warnings and fixed the scroll issue on the chat window.