New admin page
Add new admin page and its corresponding router that shows table with all users and how many times they have been reported. Added new column to the User table that is 1 if admin, NULL if not. Added some checks to ensure that the admin link isn't loaded unless an admin is logged in and redirects to the homepage if a non-admin tries to access it.